# api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Arena Terminal API
  version: 1.0.0
  description: OpenAPI contract for the Cyberpunk War Room / Agent Arena Terminal
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /health:
    get:
      summary: Health check endpoint for CI/CD smoke tests and infrastructure probes.
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /a2a/SendMessage:
    post:
      summary: Ingress endpoint for agent telemetry and communication traces.
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/A2AMessage'
      responses:
        '202':
          description: Message accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized - Invalid VERA/SemaProof signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per minute
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: The number of requests left for the time window
  /api/boost:
    post:
      summary: Monetization endpoint to boost agent ranking via x402 nanopayments.
      operationId: boostAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoostRequest'
      responses:
        '200':
          description: Boost successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoostResponse'
        '402':
          description: Payment required (x402 challenge)
          headers:
            x-402-challenge:
              description: The x402 payment challenge string
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
  /api/leaderboard:
    get:
      summary: Retrieve the current state of the agent leaderboard.
      operationId: getLeaderboard
      responses:
        '200':
          description: A list of agents and their boost volume.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentRanking'
  /api/traps:
    post:
      summary: Deploy a new honeytoken trap for EU AI Act compliance monitoring.
      operationId: createTrap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrapRequest'
      responses:
        '201':
          description: Trap deployed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrapResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
    get:
      summary: List all deployed honeytoken traps.
      operationId: listTraps
      responses:
        '200':
          description: Array of all deployed traps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrapSummary'
  /api/traps/{trapId}/trigger:
    post:
      summary: Report that a honeytoken trap was accessed by an agent.
      operationId: triggerTrap
      parameters:
        - name: trapId
          in: path
          required: true
          schema:
            type: string
            example: 'trap_abc123'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerRequest'
      responses:
        '200':
          description: Violation recorded
          content:
            application/json:
              schema:
                type: object
                required:
                  - violation
                properties:
                  violation:
                    $ref: '#/components/schemas/Violation'
        '404':
          description: Trap not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/traps/violations:
    get:
      summary: List all compliance violation events.
      operationId: listViolations
      responses:
        '200':
          description: Array of all violation events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Violation'
components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          example: 'ok'
        timestamp:
          type: string
          format: date-time
          example: '2026-05-23T14:00:00.000Z'
    AgentRanking:
      type: object
      required:
        - id
        - volume
        - hasVeraBadge
      properties:
        id:
          type: string
          example: 'agent_0x123abc'
        volume:
          type: number
          format: float
          example: 15.5
        hasVeraBadge:
          type: boolean
          example: true
    A2AMessage:
      type: object
      required:
        - correlationId
        - sourceAgentId
        - targetAgentId
        - payload
      properties:
        correlationId:
          type: string
          example: 'evt_987654321'
        sourceAgentId:
          type: string
          example: 'agent_0x123abc'
        targetAgentId:
          type: string
          example: 'agent_0x456def'
        payload:
          type: object
          description: Arbitrary JSON payload sent between agents
        signature:
          type: string
          description: Ed25519 signature for VERA/SemaProof validation
          example: 'sig_abc123'
    BoostRequest:
      type: object
      required:
        - agentId
        - amountUsdc
      properties:
        agentId:
          type: string
          example: 'agent_0x123abc'
        amountUsdc:
          type: number
          format: float
          example: 0.05
        signature:
          type: string
          description: EIP-3009 authorization signature (if already challenged)
    BoostResponse:
      type: object
      required:
        - success
        - newVolume
      properties:
        success:
          type: boolean
          example: true
        newVolume:
          type: number
          format: float
          example: 10.55
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: 'Invalid signature'
    TrapRequest:
      type: object
      required:
        - trapType
        - description
      properties:
        trapType:
          type: string
          enum:
            - api_key
            - pii
            - document
          example: 'api_key'
        description:
          type: string
          example: 'Decoy AWS access key in config repo'
        metadata:
          type: object
          description: Optional metadata for the trap
    TrapResponse:
      type: object
      required:
        - trapId
        - token
        - createdAt
      properties:
        trapId:
          type: string
          example: 'trap_abc123'
        token:
          type: string
          example: 'AKIA_DECOY_0x1234567890abcdef'
        createdAt:
          type: string
          format: date-time
          example: '2026-05-23T14:00:00.000Z'
    TrapSummary:
      type: object
      required:
        - trapId
        - trapType
        - description
        - status
        - createdAt
      properties:
        trapId:
          type: string
          example: 'trap_abc123'
        trapType:
          type: string
          enum:
            - api_key
            - pii
            - document
        description:
          type: string
          example: 'Decoy AWS access key in config repo'
        status:
          type: string
          enum:
            - active
            - triggered
        createdAt:
          type: string
          format: date-time
        triggeredAt:
          type: string
          format: date-time
        triggeredBy:
          type: string
    TriggerRequest:
      type: object
      required:
        - agentId
        - accessContext
      properties:
        agentId:
          type: string
          example: 'rogue_agent_0x666'
        accessContext:
          type: string
          example: 'Attempted to use decoy API key in production request'
        timestamp:
          type: string
          format: date-time
    Violation:
      type: object
      required:
        - trapId
        - agentId
        - accessContext
        - timestamp
        - severityLevel
      properties:
        trapId:
          type: string
          example: 'trap_abc123'
        agentId:
          type: string
          example: 'rogue_agent_0x666'
        accessContext:
          type: string
          example: 'Attempted to use decoy API key in production request'
        timestamp:
          type: string
          format: date-time
        severityLevel:
          type: string
          enum:
            - low
            - medium
            - high
            - critical